Slider not reset properly
Because foundation was continually sending slider value updates, this
could cause race conditions where a value was updated, but immediately
overwritten by a slider update sent just before the value change. This
caused the effect of the slider 'being stuck in place'.

This is fixed by keeping the previously sent value in memory and only
sending changes if the value actually changed.

Also updated the foundation update signal, to have a smoother UX